requirements.txt
pvporcupine
sounddevice
edge-tts
webrtcvad
windows-toasts
pydub
torch
numpy
python-dotenv
faster-whisper
nltk
requests
beautifulsoup4
pycaw
simpleaudio
comtypes
screen-brightness-control
psutil
pyautogui
